from polymerelements to PolymerElements
This is a similar pull request like #46 which I closed. The polymer team did the 2.0-preview merge a while ago and this pull request is against the master branch.

The rational for this change:

This pull request wants to make this Polymer element consistent with the majority of other Polymer elements. The uppercase version "PolymerElements" and "Polymer" is closer to real name of the github project name, like presented in the git URL.

The use of mixed case does not seem to have an effect on bower and JavaScript projects. But other languages like Java are more picky and would benefit from consistency.

I'm mostly concerned with the bower "Main" dependencies, but as I have changed this file already I also fixed this in the devDependencies

This pull request is a manual follow up of PolymerLabs/tedium#47 and PolymerLabs/tedium#48 which try to do this in an automated way, but are stuck.